finish reorder api

+++ app/controllers/list.controller.js
@@ -1,4 +1,5 @@
 const db = require('../models');
+const { getBoardWithList, getBoardWithEverything } = require('./helpers');
 const Board = db.board;
 const List = db.list;
 const Task = db.task;
@@ -6,6 +7,10 @@ const Task = db.task;
 const Op = db.Sequelize.Op;
 
 exports.createList = async (req, res) => {
+  const board = await getBoardWithList(req.params.id);
+  if (!board) {
+    return res.status(404).send({ message: 'Board Not Found' });
+  }
   let list = await List.create({
     BoardId: req.params.id,
     name: req.body.name,
@@ -16,6 +21,10 @@ exports.createList = async (req, res) => {
 };
 
 exports.getLists = async (req, res) => {
+  const board = await getBoardWithList(req.params.id);
+  if (!board) {
+    return res.status(404).send({ message: 'Board Not Found' });
+  }
   let lists = await List.findAll({
     where: {
       BoardId: req.params.id,
@@ -27,6 +36,10 @@ exports.getLists = async (req, res) => {
 };
 
 exports.renameList = async (req, res) => {
+  const board = await getBoardWithList(req.params.id);
+  if (!board) {
+    return res.status(404).send({ message: 'Board Not Found' });
+  }
   let list = await List.update(
     { name: req.body.name },
     {
@@ -41,6 +54,10 @@ exports.renameList = async (req, res) => {
 };
 
 exports.deleteList = async (req, res) => {
+  const board = await getBoardWithList(req.params.id);
+  if (!board) {
+    return res.status(404).send({ message: 'Board Not Found' });
+  }
   try {
     await List.destroy({
       where: {
@@ -50,5 +67,38 @@ exports.deleteList = async (req, res) => {
     res.status(200);
   } catch (error) {
     res.status(500).send({ message: error });
+  }
+};
+
+///board/:id/list/:list_id/reorder
+exports.reorderList = async (req, res) => {
+  try {
+    let board = await getBoardWithList(req.params.id);
+    if (!board) {
+      res.status(404).send({ message: 'Board not found.' });
+    }
+
+    let list = await List.findByPk(req.params.list_id);
+    if (!list) {
+      res.status(404).send({ message: 'List not found.' });
+    }
+
+    const listIds = board.Lists.map((list) => list.id);
+    if (!listIds.includes(list.id)) {
+      res.status(404).send({ message: 'List does not belong to board.' });
+    }
+
+    const reorderNum = req.body.orderNum;
+    board.Lists.splice(listIds.indexOf(list.id), 1);
+    board.Lists.splice(reorderNum, 0, list);
+
+    //update all list's order
+    for (let i of board.Lists) {
+      await i.update({ order: board.Lists.indexOf(i) + 1 });
+    }
+    board = await getBoardWithEverything(req.userId);
+    res.status(200).send({ board });
+  } catch (error) {
+    res.status(500).send({ error });
   }
 };
